TMF629/requirements.txt
fastapi
uvicorn
httpx
fastapi-mcp
pyyaml
